+++ config/packages.nix
@@ -20,7 +20,7 @@
     slurp
     nwg-look
     rofi
-    wofi
+    #wofi
     waybar
     matugen
     wl-clipboard
@@ -32,7 +32,9 @@
     qt6.qtmultimedia
 
     # Add your packages here
+    alacritty
     alejandra
+    appimage-run
     #assaultcube
     atop
     bat
@@ -48,8 +50,10 @@
     #discord
     #discord-canary
     fastfetch
+    fd # better find used by emacs, etc
     foot
     #gajim # Japper XMPP client
+    gearlever # manage app iamges
     gcc
     ghostty
     git
@@ -68,12 +72,13 @@
     mesa-demos # needed for inxi
     mpv # video player
     nemo-with-extensions # nemo file mgr
+    neovide # gui for neovim
     ncdu # show diskusage
     nh # Nix Helper
     nixd # nix lsp
     #nvtopPackages.full # nvtop all GPUs
     onefetch # git repo fetch
-    onlyoffice-desktopeditors
+    #onlyoffice-desktopeditors
     pciutils
     ripgrep
     #rustup
@@ -82,7 +87,7 @@
     ttop # resource monitor
     ugrep
     vlc # Video Player
-    warp-terminal # AI and Terminal
+    #warp-terminal # AI and Terminal
     wezterm # Terminal
     wget
     zenith # Btop/htop/bottom style monitor